Better support for complex inheritance
Given the case when `A` → `B` → `C`, and `A` has property `id`, which is overridden by the property `id`, which in its turn is overridden in `C` with the property `id` the generated code was invalid before, it generated code like

```kotlin
sealed class A(
  open val id: Int
)

open class B(
  open val id: Int // Incorrect, should have the override modifier
): A()

data class C(
  val id: Int // Incorrect, should have the override modifier
): B()
```

We can observe 2 issues in here:
1. `B` actually should be `sealed`: The hierarchy is close it doesn't make sense to allow end-users to extend it
2. `id` in both `B` and `C` should have the `override` modifier

This commit addresses both issues, introducing the new entity: `SchemaComponentDescriptor`. This entity store name, properties, and references to ancestors of each component.

Given this schema we always know if we override anything for any given property or not.